mern-blog: add comments schema and routes

+++ backend/routes/blogPostRoutes.js
@@ -1,6 +1,7 @@
 import express from 'express';
 import BlogPost from '../models/BlogPost.js';
 import cloudinaryUploader from '../config/cloudinaryConfig.js';
+import { v2 as cloudinary } from 'cloudinary';
 import { sendEmail } from '../services/emailService.js';
 // import controlloMail from '../middlewares/controlloMail.js'; // NON USARE - SOLO PER DIDATTICA - MIDDLEWARE (commentato)
 
@@ -11,7 +12,6 @@ const router = express.Router();
 // GET /blogPosts
 router.get('/', async (req, res) => {
   try {
-
     // Filter blog posts by title (case-insensitive research)
     let query = {};
     if (req.query.title) {
@@ -27,7 +27,7 @@ router.get('/', async (req, res) => {
   }
 });
 
-// GET /blogPosts/<blogPost._id>
+// GET /blogPosts/:id
 router.get('/:id', async (req, res) => {
   try {
     // Find blog post by id in MongoDB
@@ -76,7 +76,7 @@ router.post('/', cloudinaryUploader.single('cover'), async (req, res) => {
   }
 });
 
-// PUT /blogPosts/<blogPost._id>
+// PUT /blogPosts/:id
 router.put('/:id', async (req, res) => {
   try {
     // Find and update specific blog post in MongoDB
@@ -97,9 +97,23 @@ router.put('/:id', async (req, res) => {
   }
 });
 
-// DELETE /blogPosts/<blogPost._id>
+// DELETE /blogPosts/:id
 router.delete('/:id', async (req, res) => {
   try {
+    
+    // Find post by id
+    const post = await BlogPost.findById(req.params.id);
+
+    if (!post) {
+      return res.status(404).json({ message: 'Blog post not found' });
+    }
+    
+    // Pull cloudinary cover public id
+    const publicId = `blog_covers/${post.cover.split('/').pop().split('.')[0]}`
+    
+    // Delete cloudinary img
+    await cloudinary.uploader.destroy(publicId);
+
     // Find and delete specific author in MongoDB
     const deletedBlogPost = await BlogPost.findByIdAndDelete(req.params.id);
 
@@ -114,7 +128,7 @@ router.delete('/:id', async (req, res) => {
   }
 });
 
-// PATCH /blogPosts/<blogPost._id>/cover
+// PATCH /blogPosts/:id/cover
 router.patch('/:id/cover', cloudinaryUploader.single('cover'), async (req, res) => {
   try {
     if(!req.file) {
@@ -138,6 +152,129 @@ router.patch('/:id/cover', cloudinaryUploader.single('cover'), async (req, res) 
 
   } catch (err) {
     res.status(400).json({ message: err.message });
+  }
+});
+
+// GET /blogPosts/:id/comments
+router.get('/:id/comments', async (req, res) => {
+  try {
+    const post = await BlogPost.findById(req.params.id);
+
+    if (!post) {
+      return res.status(404).json({ message: 'Blog post not found' })
+    }
+
+    res.json(post.comments);
+
+  } catch (err) {
+    res.status(500).json({ message: err.message });
+  }
+});
+
+// GET /blogPosts/:id/comments/:commentId
+router.get('/:id/comments/:commentId', async (req, res) => {
+  try {
+    const post = await BlogPost.findById(req.params.id);
+
+    if (!post) {
+      return res.status(404).json({ message: 'Blog post not found' })
+    }
+
+    const comment = post.comments.id(req.params.commentId);
+
+    if (!comment) {
+      return res.status(404).json({ message: 'Comment not found' })
+    }
+
+    res.json(comment);
+
+  } catch (err) {
+    res.status(500).json({ message: err.message });
+  }
+});
+
+// POST /blogPosts/:id/comments
+router.post('/:id/comments', async (req, res) => {
+  try {
+    const post = await BlogPost.findById(req.params.id);
+
+    if (!post) {
+      return res.status(404).json({ message: 'Blog post not found' })
+    }
+
+    const newComment = {
+      name: req.body.name,
+      email: req.body.email,
+      content: req.body.content,
+    }
+
+    // Add comments
+    post.comments.push(newComment);
+
+    // Save comment on MongoDB
+    await post.save();
+
+    res.status(201).json(comment);
+
+  } catch (err) {
+    res.status(400).json({ message: err.message });
+  }
+});
+
+// PATCH /blogPosts/:id/comments/:commentId
+router.patch('/:id/comments/:commentId', async (req, res) => {
+  try {
+    const post = await BlogPost.findById(req.params.id);
+
+    if (!post) {
+      return res.status(404).json({ message: 'Blog post not found' })
+    }
+
+    const comment = post.comments.id(req.params.commentId);
+
+    if (!comment) {
+      return res.status(404).json({ message: 'Comment not found' })
+    }
+
+    // TODO:
+    comment.content = req.body.content;
+
+    // Save comment on MongoDB
+    await post.save();
+
+    res.json(comment);
+
+  } catch (err) {
+    res.status(400).json({ message: err.message });
+  }
+});
+
+// DELETE /blogPosts/:id/comments/:commentId
+router.delete('/:id/comments/:commentId', async (req, res) => {
+  try {
+    const post = await BlogPost.findById(req.params.id);
+
+    if (!post) {
+      return res.status(404).json({ message: 'Blog post not found' })
+    }
+
+    const comment = post.comments.id(req.params.commentId);
+
+    if (!comment) {
+      return res.status(404).json({ message: 'Comment not found' })
+    }
+
+    // delete comment
+    comment.remove();
+
+    // Save post on MongoDB
+    await post.save();
+
+    // Send message
+    res.json({ message: 'comment deleted' });
+
+  } catch (err) {
+    res.status(500).json({ message: err.message });
   }
 });
